end date refactoring + add modal form for deleting and adding organization

+++ AngularApi/angular-sapient/src/app/orgList/orgList.component.ts
@@ -3,6 +3,7 @@ import {OrganizationService} from "../shared/services/organization.service";
 import {HttpInterceptor} from "@angular/common/http";
 import {SearchPipe} from "../shared/search.pipe";
 import {Organization} from "../shared/interfaces";
+import Swal from "sweetalert2";
 
 @Component({
   selector: 'app-orgList',
@@ -10,21 +11,49 @@ import {Organization} from "../shared/interfaces";
   styleUrls: ['./orgList.component.scss']
 })
 
-export class OrgListComponent  implements OnInit{
+export class OrgListComponent  implements OnInit {
   lstOrgs = []
   searchStr = ''
-  counter: number = 1
+
+
   constructor(private organizationService: OrganizationService) {
   }
-  ngOnInit(){
-    this.lstOrgs = JSON.parse(<string>localStorage.getItem('Organizations'))
-    console.log(this.searchStr)
+
+  ngOnInit() {
+    this.lstOrgs = JSON.parse(<string>localStorage.getItem('Organizations'));
+
+
+    let orgDate;
+    let arr = this.lstOrgs
+    for(let obj of arr){
+      let littleArr = Object.values(obj)
+      let fullDate = littleArr[4]
+
+
+      // orgDate = fullDate.getFullYear()
+    }
+    // console.log(arr)
+    // console.log(this.searchStr)
 
   }
-  onDelete(id: any){
-    let results:any = this.lstOrgs.slice(id)
-    localStorage.setItem('Organizations', JSON.stringify(results))
 
-    this.lstOrgs = results
+  onDelete(id: any) {
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Delete',
+      confirmButtonColor: '#5ea14a',
+      cancelButtonColor: '#d33',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire('Deleted!', '', 'success')
+
+        let results:any = this.lstOrgs.slice(id)
+        localStorage.setItem('Organizations', JSON.stringify(results))
+        this.lstOrgs = results
+      }
+    })
   }
 }
